[WIP] Naming refactoring
