+++ src/Entities/Method/Delete.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Somecode\OpenApi\Entities\Methods;
+namespace Somecode\OpenApi\Entities\Method;
 
 use Somecode\OpenApi\Enums\RequestMethod;
 
-class Post extends Method
+class Delete extends Method
 {
     public function method(): RequestMethod
     {
-        return RequestMethod::POST;
+        return RequestMethod::DELETE;
     }
 }
