+++ src/Entities/Method/Method.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Somecode\OpenApi\Entities\Methods;
+namespace Somecode\OpenApi\Entities\Method;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Somecode\OpenApi\Entities\Parameters\Parameter;
+use Somecode\OpenApi\Entities\Parameter\Parameter;
 use Somecode\OpenApi\Enums\RequestMethod;
 
 abstract class Method
