+++ src/Entities/Method/Put.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Somecode\OpenApi\Entities\Methods;
+namespace Somecode\OpenApi\Entities\Method;
 
 use Somecode\OpenApi\Enums\RequestMethod;
 
-class Patch extends Method
+class Put extends Method
 {
     public function method(): RequestMethod
     {
-        return RequestMethod::PATCH;
+        return RequestMethod::PUT;
     }
 }
